backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.41
pydantic==1.10.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==0.21.1
email-validator==2.0.0 